Django/requirements.txt
requests
django-cors-headers
channels
